extract Endpoint stuff to its own file
